Introduce MQTT
This implements the core functionality of the library. Using the rumqttc
 crate the library can publish and subscribe to MQTT.

The configuration is default-initialized but already stored with and
loaded from M*Player. It's possible to change it by modifying the
M*Player project file. A connection is only established once the plugin
configuration took place. While the user won't see anything this still
seems less surprising than having the plugin immediately connect to
localhost when it got loaded.

All state of the library is stored in static global `Option` variables
behind a `Mutex` or `RwLock`. There can only ever be a single instance
running and no other scope exists where data can persist between
function calls. The `Option` allow storing values which can't be
initialized statically. While most calls should come from the same
thread at least the MQTT thread is an exception. To not rely on unsafe
code blocks a `Mutex` or `RwLock` is used for all those variables. While
the PosionError of the mutex is ignored with `unwrap` the `Option` being
`None` is handled with `expect` to document the assumption why it should
be `Some` at that point.

The types describing M*Player's plugin interface were extracted into
their own file. In `mstarInit` the provided struct is now also stored to
be able to call those functions.