Add configuration dialog
The user interface is built using Slint. Of the three evaluated crates
this is the only one that worked.

* `eframe` crashed on Linux on second launch and on Windows on first
* launch. `iced` crashed on Linux with: "Creating EventLoop multiple
* times is not supported."

With slint launching the dialog multiple times works flawlessly on Linux
and Windows.

The dialog allows the user to configure the plugin. When apply is
pressed a new MQTT connection is always reestablished. This is simpler
to implement and might help in error situations. When no configuration
exists yet the default values are prefilled for convenience.