Add integration test
As mentioned upstream[^1] the integration tests aren't able to use
cdylib crate types so "lib" was added to the list of crate types.

To not having to repeat the plugin interface types for Rust in the
integration test they were made public in the library.

The integration test needs a running MQTT server to which it will
connect. That way the full functionality (except for mstarConfigure with
its UI) can be tested.

[^1]: https://github.com/rust-lang/cargo/issues/6659#issuecomment-463335095