Fix crash when plugin wasn't configured yet
The intention was that the plugin won't crash when it gets loaded and
M*Player calls its functions but it wasn't configured yet to connect to
any MQTT server.

The `startup_shutdown` test didn't catch that reliably since it runs
the `configuration` test on another thread in the same process. This
meant the configuration could actually have been set, thus making the
test succeed although the plugin would have crashed.

The test will be fixed in a later commit. This only fixes the wrong
behavior.