Make startup_shutdown an integration test
As described in the previous commit this test needs to run in its own
process. Therefore it can't be a unit test but must be its own
integration test.

To prevent code duplication and also make use of the mocks reporting
unexpected calls the code was also refactored to make the mocks
available to both integration tests.