Fix crash when saving before configuring
The MStarPlayer crashed when the project was saved without having
configured the plugin. This of course must not happen.

Fixed by not expecting `CONFIG` to actually contain a value. If there is
none an empty string is returned as the configuration. When loading the
configuration this then tears down any existing MQTT connection and
resets the configuration. Effective putting the plugin in the same state
as when it was initialized without any configuration.