Handle MQTT reconnects
Until now when an MQTT error occurred the whole sending logic shut down
which means the plugin would no longer do anything.

This commit changes the behavior to continousely try to reconnect.

The usual implementation of this would be to subscribe on `ConAck` as
suggested upstream[^1]. This will fail / block when the plugin already
queued too many packets which can quickly happen when some players are
playing. Therefore a more unusual approach was taken.

An outer loop was introduced which starts with subscribing to the topic
and then reading all messages until an error occurs. In that case the
end of the loop waits a bit and removes all accumulated messages possibly
being queued for publishing by the plugin before enqueueing the request
to subscribe.

This makes it practically impossible that the queue is filled up before
the subscription request is enqueued. When reconnecting the queued
request to subscribe to the topic is then executed.

Clearing the queued packets only became possible in rumqttc 0.24, so it
was updated to that version.

Also removed the unnecessary `mut` on the `client` variable.

[^1]: https://github.com/bytebeamio/rumqtt/issues/250#issuecomment-783296978